make it easier to port to `Core.Compiler`
